feat(Cards) : 플레이어의 카드 리스트를 일급 컬렉션으로 수정
